Change URL Header from HTTPS to SSH
Note that there may be systems in which we cannot use SSH - we'd better
check `global` or `system` settings to check if Git uses SSH as its
default connection method.

Alternatively, we could just try cloning with SSH first, and if it
returns an error, we then move on to HTTPS as a backup. If we can add a
config flag to predetermine SSH usage, that could also be useful.